+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Support for validators with compound names by restoring ful case sensitivity
   to error messages. (#805)
 - new event 'form:submit.parsley' fired before a form is submitted.
+- remote cache now cleared after form submission (#813)
 
 ## 2.1.0-rc1